Support enabling the addon automatically.
This is partially to support travis because I figured out how
to programmatically enable an addon but not how to save the user
settings. I actually think that just enabling the addon when it
is needed is actually cleaner. That enables the uesr to use blender
without needing to enable the addon and let the build enable it
when it needs it.