Add batching API and reimplement CLI around it.
Fixes #2.

This adds a worker class called BabylonjsBlenderWorker which is just
a transform stream which takes job objects and emits them on completion.
Each worker is capable of processing one job at a time.

The CLI was reimplemented in terms of the worker API. It creates a pool
of workers whose size is controlled with the help of gulp-lock.

The tests were mostly rewritten as mocha, particularly the new tests
of the API.