Implement Windows-compatible worker reuse mechanism.
Now when a worker is used, the actual blender process is reused.
To communicate with blender instances, the module talks in a weird
protocol I should not have made up over stdin/stdout (because inheriting
other file descriptors does not work with Python on Windows; it seems
they made stdin/stdout/stderr a special case). The result is much faster
runs when there are multiple scenes because the startup time of Blender
is sufficiently long.

You should be able to disable this behavior by setting the
NODEJS_BABYLONJS_BLENDER_NOSTDIOCHAT environment variable.