Add node-babylonjs-blender-install command.
This command automatically downloads and installs a specific version
of the exporter. It also installs any Blender addons specified as
arguments. This is added to make writing the CI scripts easier by
consolidating the code between appveyor.yml and .travis.yml and
may also be helpful for writing CI for modules depending on this.