Fix support for specifying file in addon installer.
The blender addon/plugin installer helper,
node-babylonjs-blender-install, claims to allow you to pass in the path
to a .zip for a blender extension and install it for you. That functionality
apparently never worked. This fixes it, allowing you to use it to install
new versions of the plugin or arbitrary plugins as you need.